ai-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
